refactor: split popover actions
